requirements.txt
streamlit>=1.20
openai>=0.27.0
python-dateutil
pandas
